removing circular dependency
