+++ src/components/FlightFlagger/FlightFlaggerHighlightChip.stories.tsx
@@ -1,0 +1,17 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { FlightHighlightChip as FlightHighlightChipComponent } from './FlightFlaggerHighlightChip';
+
+const meta: Meta<typeof FlightHighlightChipComponent> = {
+  title: "DRT Components/Flight Flagger",
+  component: FlightHighlightChipComponent,
+};
+
+export default meta;
+type Story = StoryObj<typeof FlightHighlightChipComponent>;
+
+export const HighlightChip: Story = {
+  args: {
+    text: 'Highlighted info',
+  },
+};
